+++ CHANGELOG.md
@@ -1,13 +1,14 @@
 # Change Log
 All notable changes to this project will be documented in this file.
 
-## [1.0.1] - 2019-03-17 {dev}
+## [1.0.1] - 2019-03-28 {dev}
 ### Added
 - New {pagepath} tag.
 - New {editingtoggle} tag.
 - New {idnumber} tag.
 - New {fa...} tag (for FontAwesome).
 - New {glyphicon...} tag (for Glyphicons).
+- New {sessionid} and %7Bsessionid%7D tags.
 ### Updated
 - Fixed {categories} filter code compatibility with Moodle 2.7 to 3.5.